requirements.txt
Django==5.0.4
requests==2.32.0
scikit-learn==1.5.0
numpy==1.26.4
scipy==1.13.0
python-decouple==3.8
dj-database-url==2.2.0
psycopg[binary]==3.2.1
redis==5.0.7
celery==5.4.0
django-crispy-forms==2.3
crispy-bootstrap5==2024.2
gunicorn==22.0.0
whitenoise==6.7.0
